Move public headers to include/firebird with wrappers in include for legacy files.
