+++ src/include/ib_util.h
@@ -1,4 +1,8 @@
 /*
+ *	PROGRAM:	UDF and Blob filter Utilities library
+ *	MODULE:		ib_util.h
+ *	DESCRIPTION:	Prototype header file for ib_util.c
+ *
  * The contents of this file are subject to the Interbase Public
  * License Version 1.0 (the "License"); you may not use this file
  * except in compliance with the License. You may obtain a copy
@@ -8,6 +12,13 @@
  * "AS IS" basis, WITHOUT WARRANTY OF ANY KIND, either express
  * or implied. See the License for the specific language governing
  * rights and limitations under the License.
+ *
+ * The Original Code was created by Inprise Corporation
+ * and its predecessors. Portions created by Inprise Corporation are
+ * Copyright (C) Inprise Corporation.
+ *
+ * All Rights Reserved.
+ * Contributor(s): ______________________________________.
  */
 
-#include "firebird/iberror.h"
+#include "firebird/ib_util.h"
